refactor get interactions into a shared file
